requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cffi==1.15.0
coverage==6.1.2
cssselect2==0.4.1
dj-database-url==0.5.0
Django==3.2.7
django-braces==1.14.0
django-crispy-forms==1.12.0
django-environ==0.7.0
django-heroku==0.3.1
django-parsley==0.7
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
fonttools==4.27.1
gunicorn==20.1.0
html5lib==1.1
idna==3.3
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
pydyf==0.1.2
pyphen==0.11.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
validate-email==1.3
weasyprint==53.3
webencodings==0.5.1
whitenoise==5.3.0
xlwt==1.3.0
zopfli==0.1.9
